refactor inner rekey test
